// src/shared/api/base/generated-types.ag.ts
export enum PSupportedLanguageCodes {
	de = 'de',
	en = 'en',
}

/**
 * All country-codes that are available in our App (last two chars from LOCALE_ID).
 */
export enum PSupportedCountryCodes {
	DE = 'DE',
	AT = 'AT',
	CH = 'CH',
	NL = 'NL',
	BE = 'BE',
	GB = 'GB',
	CZ = 'CZ',
	SE = 'SE',
	LU = 'LU',
}